CIEnvironmentValues refactor
