+++ tracer/src/Datadog.Trace/Ci/Agent/CIVisibilityProtocolWriter.cs
@@ -81,7 +81,7 @@ namespace Datadog.Trace.Ci.Agent
 
             // Concurrency Level is a simple algorithm where we select a number between 1 and 4 depending on the number of Logical Processor Count
             // To scale the number of senders with a hard limit.
-            var concurrencyLevel = concurrency ?? Math.Min(Math.Max(Environment.ProcessorCount / 2, 1), 4);
+            var concurrencyLevel = concurrency ?? Math.Min(Math.Max(System.Environment.ProcessorCount / 2, 1), 4);
             _buffersArray = new Buffers[concurrencyLevel];
             for (var i = 0; i < _buffersArray.Length; i++)
             {
