+++ tracer/src/Datadog.Trace/Ci/CIVisibility.cs
@@ -14,6 +14,7 @@ using Datadog.Trace.Agent;
 using Datadog.Trace.Agent.DiscoveryService;
 using Datadog.Trace.Agent.Transports;
 using Datadog.Trace.Ci.Configuration;
+using Datadog.Trace.Ci.Environment;
 using Datadog.Trace.Configuration;
 using Datadog.Trace.Logging;
 using Datadog.Trace.Pdb;
@@ -212,12 +213,12 @@ namespace Datadog.Trace.Ci
                 AsyncUtil.RunSync(() => FlushAsync(), cts.Token);
                 if (cts.IsCancellationRequested)
                 {
-                    Log.Error("Timeout occurred when flushing spans.{NewLine}{StackTrace}", Environment.NewLine, Environment.StackTrace);
+                    Log.Error("Timeout occurred when flushing spans.{NewLine}{StackTrace}", System.Environment.NewLine, System.Environment.StackTrace);
                 }
             }
             catch (TaskCanceledException)
             {
-                Log.Error("Timeout occurred when flushing spans.{NewLine}{StackTrace}", Environment.NewLine, Environment.StackTrace);
+                Log.Error("Timeout occurred when flushing spans.{NewLine}{StackTrace}", System.Environment.NewLine, System.Environment.StackTrace);
             }
             finally
             {
@@ -459,7 +460,7 @@ namespace Datadog.Trace.Ci
                         break;
                 }
 
-                return Environment.OSVersion.VersionString;
+                return System.Environment.OSVersion.VersionString;
             }
         }
 
@@ -492,9 +493,9 @@ namespace Datadog.Trace.Ci
                 {
                     processName ??= GetProcessName();
                     // When is enabled by configuration we only enable it to the testhost child process if the process name is dotnet.
-                    if (processName.Equals("dotnet", StringComparison.OrdinalIgnoreCase) && Environment.CommandLine.IndexOf("testhost.dll", StringComparison.OrdinalIgnoreCase) == -1)
+                    if (processName.Equals("dotnet", StringComparison.OrdinalIgnoreCase) && System.Environment.CommandLine.IndexOf("testhost.dll", StringComparison.OrdinalIgnoreCase) == -1)
                     {
-                        Log.Information("CI Visibility disabled because the process name is 'dotnet' but the commandline doesn't contain 'testhost.dll': {Cmdline}", Environment.CommandLine);
+                        Log.Information("CI Visibility disabled because the process name is 'dotnet' but the commandline doesn't contain 'testhost.dll': {Cmdline}", System.Environment.CommandLine);
                         return false;
                     }
 
@@ -535,7 +536,7 @@ namespace Datadog.Trace.Ci
                 try
                 {
                     // Set the configuration key to propagate the configuration to child processes.
-                    Environment.SetEnvironmentVariable(ConfigurationKeys.CIVisibility.Enabled, "1", EnvironmentVariableTarget.Process);
+                    System.Environment.SetEnvironmentVariable(ConfigurationKeys.CIVisibility.Enabled, "1", EnvironmentVariableTarget.Process);
                 }
                 catch
                 {
