+++ tracer/src/Datadog.Trace/Configuration/GitMetadata.cs
@@ -4,13 +4,13 @@
 // </copyright>
 
 #nullable enable
-using Datadog.Trace.Ci;
+using Datadog.Trace.Ci.Environment;
 
 namespace Datadog.Trace.Configuration;
 
 internal class GitMetadata
 {
-    public static readonly GitMetadata Empty = new GitMetadata(string.Empty, string.Empty);
+    public static readonly GitMetadata Empty = new(string.Empty, string.Empty);
 
     public GitMetadata(string commitSha, string repositoryUrl)
     {
