+++ tracer/test/Datadog.Trace.ClrProfiler.IntegrationTests/CI/MsTestV2EvpTests.cs
@@ -9,6 +9,7 @@ using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 using Datadog.Trace.Ci;
+using Datadog.Trace.Ci.Environment;
 using Datadog.Trace.Ci.Tags;
 using Datadog.Trace.Configuration;
 using Datadog.Trace.ExtensionMethods;
@@ -22,7 +23,7 @@ using Xunit.Abstractions;
 
 namespace Datadog.Trace.ClrProfiler.IntegrationTests.CI
 {
-    public class MsTestV2EvpTests : TestHelper
+    public class MsTestV2EvpTests : TestingFrameworkEvpTest
     {
         private const string TestSuiteName = "Samples.MSTestTests.TestSuite";
         private const string TestBundleName = "Samples.MSTestTests";
@@ -122,7 +123,7 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CI
                         }
                     };
 
-                    using (ProcessResult processResult = await RunDotnetTestSampleAndWaitForExit(agent, packageVersion: packageVersion, copyCIEnvironmentValues: true))
+                    using (ProcessResult processResult = await RunDotnetTestSampleAndWaitForExit(agent, packageVersion: packageVersion))
                     {
                         // Check the tests, suites and modules count
                         Assert.Equal(expectedTestCount, tests.Count);
@@ -301,174 +302,6 @@ namespace Datadog.Trace.ClrProfiler.IntegrationTests.CI
                 WriteSpans(tests);
                 throw;
             }
-        }
-
-        private static void WriteSpans(List<MockCIVisibilityTest> tests)
-        {
-            if (tests is null || tests.Count == 0)
-            {
-                return;
-            }
-
-            Console.WriteLine("***********************************");
-
-            int i = 0;
-            foreach (var test in tests)
-            {
-                Console.Write($" {i++}) ");
-                Console.Write($"TraceId={test.TraceId}, ");
-                Console.Write($"SpanId={test.SpanId}, ");
-                Console.Write($"Service={test.Service}, ");
-                Console.Write($"Name={test.Name}, ");
-                Console.Write($"Resource={test.Resource}, ");
-                Console.Write($"Type={test.Type}, ");
-                Console.Write($"Error={test.Error}");
-                Console.WriteLine();
-                Console.WriteLine($"   Tags=");
-                foreach (var kv in test.Meta)
-                {
-                    Console.WriteLine($"       => {kv.Key} = {kv.Value}");
-                }
-
-                Console.WriteLine();
-            }
-
-            Console.WriteLine("***********************************");
-        }
-
-        private static void AssertTargetSpanAnyOf(MockCIVisibilityTest targetTest, string key, params string[] values)
-        {
-            string actualValue = targetTest.Meta[key];
-            Assert.Contains(actualValue, values);
-            targetTest.Meta.Remove(key);
-        }
-
-        private static void AssertTargetSpanEqual(MockCIVisibilityTest targetTest, string key, string value)
-        {
-            Assert.Equal(value, targetTest.Meta[key]);
-            targetTest.Meta.Remove(key);
-        }
-
-        private static void AssertTargetSpanExists(MockCIVisibilityTest targetTest, string key)
-        {
-            Assert.True(targetTest.Meta.ContainsKey(key));
-            targetTest.Meta.Remove(key);
-        }
-
-        private static void AssertTargetSpanContains(MockCIVisibilityTest targetTest, string key, string value)
-        {
-            Assert.Contains(value, targetTest.Meta[key]);
-            targetTest.Meta.Remove(key);
-        }
-
-        private static void CheckCIEnvironmentValuesDecoration(MockCIVisibilityTest targetTest, string repository, string branch, string commitSha)
-        {
-            var context = new SpanContext(parent: null, traceContext: null, serviceName: null);
-            var span = new Span(context, DateTimeOffset.UtcNow);
-            CIEnvironmentValues.Create().DecorateSpan(span);
-
-            AssertEqual(CommonTags.CIProvider);
-            AssertEqual(CommonTags.CIPipelineId);
-            AssertEqual(CommonTags.CIPipelineName);
-            AssertEqual(CommonTags.CIPipelineNumber);
-            AssertEqual(CommonTags.CIPipelineUrl);
-            AssertEqual(CommonTags.CIJobUrl);
-            AssertEqual(CommonTags.CIJobName);
-            AssertEqual(CommonTags.StageName);
-            AssertEqual(CommonTags.CIWorkspacePath);
-            AssertEqual(CommonTags.GitRepository, repository);
-            AssertEqual(CommonTags.GitCommit, commitSha);
-            AssertEqual(CommonTags.GitBranch, branch);
-            AssertEqual(CommonTags.GitTag);
-            AssertEqual(CommonTags.GitCommitAuthorName);
-            AssertEqual(CommonTags.GitCommitAuthorEmail);
-            AssertEqualDate(CommonTags.GitCommitAuthorDate);
-            AssertEqual(CommonTags.GitCommitCommitterName);
-            AssertEqual(CommonTags.GitCommitCommitterEmail);
-            AssertEqualDate(CommonTags.GitCommitCommitterDate);
-            AssertEqual(CommonTags.GitCommitMessage);
-            AssertEqual(CommonTags.BuildSourceRoot);
-
-            void AssertEqual(string key, string value = null)
-            {
-                if (value is null)
-                {
-                    if (span.GetTag(key) is { } keyValue)
-                    {
-                        Assert.Equal(keyValue, targetTest.Meta[key]);
-                        targetTest.Meta.Remove(key);
-                    }
-                }
-                else
-                {
-                    Assert.Equal(value, targetTest.Meta[key]);
-                    targetTest.Meta.Remove(key);
-                }
-            }
-
-            void AssertEqualDate(string key)
-            {
-                if (span.GetTag(key) is { } keyValue)
-                {
-                    Assert.Equal(DateTimeOffset.Parse(keyValue), DateTimeOffset.Parse(targetTest.Meta[key]));
-                    targetTest.Meta.Remove(key);
-                }
-            }
-        }
-
-        private static void CheckRuntimeValues(MockCIVisibilityTest targetTest)
-        {
-            AssertTargetSpanExists(targetTest, CommonTags.RuntimeName);
-            AssertTargetSpanExists(targetTest, CommonTags.RuntimeVersion);
-            AssertTargetSpanExists(targetTest, CommonTags.RuntimeArchitecture);
-            AssertTargetSpanExists(targetTest, CommonTags.OSArchitecture);
-            AssertTargetSpanExists(targetTest, CommonTags.OSPlatform);
-            AssertTargetSpanEqual(targetTest, CommonTags.OSVersion, CIVisibility.GetOperatingSystemVersion());
-        }
-
-        private static void CheckTraitsValues(MockCIVisibilityTest targetTest)
-        {
-            // Check the traits tag value
-            AssertTargetSpanEqual(targetTest, TestTags.Traits, "{\"Category\":[\"Category01\"],\"Compatibility\":[\"Windows\",\"Linux\"]}");
-        }
-
-        private static void CheckOriginTag(MockCIVisibilityTest targetTest)
-        {
-            // Check the test origin tag
-            AssertTargetSpanEqual(targetTest, Tags.Origin, TestTags.CIAppTestOriginName);
-        }
-
-        private static void CheckSimpleTestSpan(MockCIVisibilityTest targetTest)
-        {
-            // Check the Test Status
-            AssertTargetSpanEqual(targetTest, TestTags.Status, TestTags.StatusPass);
-        }
-
-        private static void CheckSimpleSkipFromAttributeTest(MockCIVisibilityTest targetTest)
-        {
-            // Check the Test Status
-            AssertTargetSpanEqual(targetTest, TestTags.Status, TestTags.StatusSkip);
-
-            // Check the Test skip reason
-            AssertTargetSpanEqual(targetTest, TestTags.SkipReason, "Simple skip reason");
-        }
-
-        private static void CheckSimpleErrorTest(MockCIVisibilityTest targetTest)
-        {
-            // Check the Test Status
-            AssertTargetSpanEqual(targetTest, TestTags.Status, TestTags.StatusFail);
-
-            // Check the span error flag
-            Assert.Equal(1, targetTest.Error);
-
-            // Check the error type
-            AssertTargetSpanEqual(targetTest, Tags.ErrorType, typeof(DivideByZeroException).FullName);
-
-            // Check the error stack
-            AssertTargetSpanContains(targetTest, Tags.ErrorStack, typeof(DivideByZeroException).FullName);
-
-            // Check the error message
-            AssertTargetSpanEqual(targetTest, Tags.ErrorMsg, new DivideByZeroException().Message);
         }
     }
 }
